# Accelerate cloth grid generation