## Description

I've further optimized Builder::add_cloth_grid() for enhanced performance, through:
 
1.  Fix duplicate vertex transform intrudced in last my [PR#178](https://github.com/newton-physics/newton/pull/178)
2.  Accelerate vertex transforms with NumPy vectorization

Performance improvements:
| Grid Size | Before  | After | Speedup |
|----------------------------------------|
|  200x200  |  14.6 sec.  |  1.0 sec.  |   ~x15   |

## Newton Migration Guide

Please ensure the migration guide for **warp.sim** users is up-to-date with the changes made in this MR.

- [x] The migration guide in ``docs/migration.rst`` is up-to date

## Before your PR is "Ready for review"

- [x] All commits are [signed-off](https://git-scm.com/docs/git-commit#Documentation/git-commit.txt--s) to indicate that your contribution adheres to the [Developer Certificate of Origin](https://developercertificate.org/) requirements
- [x] I understand that **GitHub** does not perform any GPU testing of this pull request
- [ ] Necessary tests have been added
- [ ] Documentation is up-to-date
- [x] Code passes formatting and linting checks with `pre-commit run -a`